Resized textures to crop out the transparent background and prevent massive hitboxes.
Added extra buffer to the bulet position based on the size of the player texture in order for the bullet to be fired from the centre after the texture resizing.

Started work on getting the keys to be able to be swapped. The controls object needs to be the same for all of the classes in order for this to take place.

Known issues:
The swapping of keybinds only works if you are pressing the key at the exact moment you click to register a new one.